refactored course management code
